requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.8.1
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.0.6
django-celery==3.3.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
httpie==1.0.3
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
kombu==3.0.37
more-itertools==7.2.0
mysql-connector==2.2.9
mysqlclient==1.4.4
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
redis==2.10.6
requests==2.22.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
wget==3.2
zipp==0.6.0
